[term] Make set_char_at() to accept u16 args instead of int

+++ include/tilck/kernel/term.h
@@ -6,7 +6,7 @@
 typedef struct {
 
    /* Main functions */
-   void (*set_char_at)(int row, int col, u16 entry);
+   void (*set_char_at)(u16 row, u16 col, u16 entry);
    void (*set_row)(int row, u16 *data, bool flush); // NOTE: set_row() can
                                                     // safely assume that it has
                                                     // been called in a FPU
